backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.152.0
requests==2.32.3
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.2
elevenlabs==1.12.0
python-multipart==0.0.12
python-dateutil==2.9.0
pytz==2024.1
apscheduler==3.10.4

